refactor(mobile): move draft orchestration into runtime
